pyproject.toml
[project]
name = "pageindex-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "jupytext>=1.19.1",
    "notebook>=7.5.4",
    "ollama>=0.6.1",
    "openai==1.101.0",
    "pageindex>=0.2.6",
    "pymupdf==1.26.4",
    "pypdf2==3.0.1",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "tiktoken==0.11.0",
]
